refactor(tidy-doc): prefactor for a second page layout (#63)
Extract the Variants Section's private specimen-instance factory into a
shared module with an optional per-cell property-override map, export the
status-badge/specimen-scene/verdict-glyph helpers for reuse, and move the
horizontal section-assembly body of the build orchestrator into its own
private routine — zero behavior change, clearing the seam for the upcoming
vertical layout (#62).
